requirements/dev.txt
black==23.12.1
pylint==3.0.3
scriv==1.5.1
twine==4.0.2
pytest==7.4.4
pytest-django==4.7.0
tox==4.11.4
